Update Disk_Cleanup.ps1
-Due to issues with deleting IIS temp files, I've added an exclude IF statement in the removal process making sure that it's not deleting anything inside of inetpub.
-Tested on servers having the issues and all was successful!

+++ Maintenance/Disk Cleanup/Powershell/Disk_Cleanup.ps1
@@ -83,8 +83,11 @@ ForEach ($folder in $folders) {
         $tempCount++
         $item = $_.FullName
         Try {
-            Remove-Item $item -Recurse -Force -ErrorAction Stop
-            #Write-Output "Deleted $item"
+            ## We've had issues with IIS if you mess with the inetpub temp so we are excluding that in this IF statement
+            If ($item -notlike '*inetpub*') {
+                Remove-Item $item -Recurse -Force -ErrorAction Stop
+                #Write-Output "Deleted $item"
+            }
         } Catch {
             $tempCount--
             #Write-Warning "Failed to delete $item"
